Changed numFreeBlocks to 32 bit

+++ includes/myfs-structs.h
@@ -56,7 +56,7 @@ struct SuperBlock {
     int32_t fatPos;
     int32_t rootPos;
     int32_t dataPos;
-    int16_t numFreeBlocks;
+    int32_t numFreeBlocks;
 };
 
 #endif /* myfs_structs_h */
